Shift creation API to be in terms of error bounds
The Bloom API was previously done up in terms of byte sizes. That
meant it was possible for a user to accidentally create a Bloom too
small to meet the error bound they expected. Now, the API has been
adjusted to error out if the error bound cannot be satisfied for
want of bytes.

The user also now has insight into the capacity and length of the
Bloom. If len > capacity then the error bound is no longer satisfied.
I may adapt the insertion return to incorporate that notion via an
error.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>