Address clippy's concerns
I intend to adjust the API -- with ambitions for a Bloom trait -- a
tad in the next commit but I figure it was time to get clippy's take
on the code so far. For the most part the code was dinged for casts,
as expected. A few style things per clippy_style were also addressed.

Not much to see here, really.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>