Introduce fuzzing for original bloom filter
This commit introduces a honggfuzz target for the original bloom
filter, changes need to stop it crashing. Most of the crashes were
masked by the QC rules meant to focus on actual tests so, uh, always
fuzz and QC software, pals.

A few more BuildErrors are now kicked out when creating a Bloom, each
in response to some found crash.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>