Bloom no longer requires explicit total_hashes
After a little bit of use I realized it was annoying for Bloom::new
to require the user to specify hashes. There is, in point of fact, a
formula to determine the optimal number of hashes to maximize the
elements that'll fit in the Bloom. This formula is now in use.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>