Introduce 'clear' method to Bloom
The 'clear' method on Bloom clears the underlying storage of the
Bloom, a necessary thing for landing in a Scalable filter.

I've started to worry that a lot of the math I've relied on for
this impl. is based on _partitioned_ bloom filters. They should
be approximately correct but it would be worth hitting the books and
doing the math out straight for Original Bloom once there is a
Partitioned Bloom available.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>