Further fleshed out the Builder for Bloom
So! It turns out my Builder approach has hit a snag with regard
to swapping out the hash_builder, specifically as I want the
final freeze method to be done in terms of the generic S but I'm
not sure how to square this with needing to create a concrete
type unless otherwise supplied.

Time to reach out to the community, see what folks do.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>