Mathy update
This commit updates the original bloom math. The user must now provide
the error bound and bit capacity and we figure out the rest from an
optimal approximation. The documentation for original bloom now also
includes the math, done in terms of the Maxima CAS. My ambition for
this project is to keep the math at the forefront of each structure's
documentation but without gatekeeping on this area of math or that.
Many structure implementations have the math implicit to the implementation
but that doesn't have to be, I don't think.

The docs themselves are still rough and not all tests pass. But, moving on
down the road.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>